feat(qemu): Expand address space to go beyond 3G
When booting secondary CPUs, the start address to boot the CPUs from is
communicated from the kernel to TF-A via PSCI.  TF-A makes sure that
address is lower than PLAT_VIRT_ADDR_SPACE_SIZE before moving forward
with the boot process.

When booting the "virt" machine with EDK2 the kernel text section is
placed at the very top of the available memory.  If that memory is
higher than 3G TF-A will refuse to boot the secondary processors because
it thinks the address is not within the platform's range.

This patch sets the physical and virtual address space known to TF-A to
a maximum of 256GB, which should be sufficient to avoid the scenario
described above.  Doing so it adjust the size of
image_info::image_max_size to accommodate the extra bits involved in
computing addresses higher than 3GB.

+++ include/export/common/bl_common_exp.h
@@ -41,7 +41,7 @@ typedef struct image_info {
 	param_header_t h;
 	uintptr_t image_base;	/* physical address of base of image */
 	uint32_t image_size;	/* bytes read from image file */
-	uint32_t image_max_size;
+	uint64_t image_max_size;
 } image_info_t;
 
 /* BL image node in the BL image execution sequence */
